requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-filter>=23.5
drf-yasg>=1.21.7
factory-boy>=3.3.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
coverage>=7.3.2
pytest>=7.4.3
pytest-django>=4.7.0